backend/requirements.txt
fastapi
uvicorn
python-dotenv
discord.py
